+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## QFeatures 1.3.3
 
-- Nothing yet
+- Improved validity checks on `AssayLinks`
+- Fixed the subsetting of `AssayLinks` to ensure consistent data
 
 ## QFeatures 1.3.2